==> notebooks/00_data.ipynb <==
This notebook analyzes data from https://github.com/HuthLab/deep-fMRI-dataset. To set up, see instructions in the `deep-fMRI-dataset` folder.

%load_ext autoreload
%autoreload 2
import datasets
import numpy as np
from os.path import join
# from ridge_utils.SemanticModel import SemanticModel
from matplotlib import pyplot as plt
from typing import List
from sklearn.linear_model import RidgeCV, LogisticRegressionCV
from sklearn.feature_extraction.text import CountVectorizer
# from feature_spaces import em_data_dir, data_dir, results_dir
from collections import defaultdict
import pandas as pd
import os
import seaborn as sns
# from fit_linear_models import get_dsets
# from ridge_utils.SemanticModel import SemanticModel
# import encoding_utils, feature_spaces
from transformers.pipelines.pt_utils import KeyDataset
# from feature_spaces import *
from transformers import pipeline
import datasets
import sys
sys.path.append('..')
import data
from tqdm import tqdm

r = defaultdict(list)
for dset in tqdm([
    # 'trec',
    'emotion', 'rotten_tomatoes', 'tweet_eval', 'sst2',
    # 'go_emotions', 'poem_sentiment', 
    # 'ethics-commonsense', 'ethics-deontology', 'ethics-justice', 'ethics-utilitarianism', 'ethics-virtue',
    'probing-subj_number', 'probing-word_content', 'probing-obj_number',
    'probing-past_present', 'probing-sentence_length', 'probing-top_constituents',
    'probing-tree_depth', 'probing-coordination_inversion', 'probing-odd_man_out',
    'probing-bigram_shift', 'moral_stories',
]):
    print(dset)
    X, y, X_test, y_test = data.get_dsets(dataset=dset, seed=1)
    values, counts = np.unique(y, return_counts=True)
    y_majority_frac_train = counts.max() / len(y)
    n_classes = values.size
    values, counts = np.unique(y_test, return_counts=True)
    y_majority_frac_test = counts.max() / len(y)
    lens_train = [len(x) for x in X]

    r['dset'].append(dset)
    r['n_train'].append(len(X))
    r['n_test'].append(len(X_test))
    r['y_majority_frac_train'].append(y_majority_frac_train)
    r['y_majority_frac_test'].append(y_majority_frac_test)
    r['len_avg_train'].append(np.mean(lens_train))
    r['len_std_train'].append(np.std(lens_train))
    r['n_classes'].append(n_classes)
    r['ex1'].append(X[0])


df = pd.DataFrame.from_dict(r)
df.to_pickle('../processed/data_table.pkl')

df = pd.read_pickle('../processed/data_table.pkl')
df.round(2)

# Look at narrative stories

import encoding_utils, feature_spaces
train_stories, test_stories, allstories = encoding_utils.get_allstories([1, 2, 3, 4, 5])

wordseqs = feature_spaces.get_story_wordseqs(allstories)

' '.join(wordseqs['sloth'].data)
