==> notebooks/06_emb.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
sys.path.append('..')
import imodelsx.llm
from transformers import AutoModel, AutoTokenizer, pipeline
from tqdm import tqdm
fit_encoding = __import__('01_fit_encoding')

# Input text
texts = ['this is a sample text that is longer than the other one',
         'the cat chased the dog']

checkpoint = 'meta-llama/Meta-Llama-3-8B'
llm = imodelsx.llm.LLMEmbs(checkpoint=checkpoint)

embs = llm(texts, layer_idx=18, batch_size=16)

inputs = llm.tokenizer_(
    texts, return_tensors='pt', padding=True).to(llm.model_.device)
hidden_states = llm.model_(**inputs).hidden_states

embs.shape
